+++ hive_organizer.py
@@ -18,28 +18,42 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import tkinter as tk
 import re
-import math
+#import math
 import os
 from tkinter.filedialog import asksaveasfilename, askopenfilename
 from tkinter import ttk
+from tkinter.colorchooser import askcolor
 from PIL import Image, ImageTk, ImageGrab
-from hive.utils import callweb, listadd, listsub, center, VerticalScrolledFrame, find
-from hive.styles import initStyle, used_colors
+from copy import deepcopy
+import json
+from hive.utils import VerticalScrolledFrame, ScrolledFrame, callweb, listadd, listsub, center, find, getPoly, rot, findMenuIndex
+from hive.styles import initStyle, used_colors, used_colors_default
+from hive.canvas import *
 
-Version = "V0.2.3.2"
-
+Version = "V0.3.alpha"
+#TODO for V0.3:
+# - Menus and Keyboard Short Cuts -> partially done
+# - color selection for all buildings -> Done
+# - Remove erased cities from Assignement List and also delete name pop-up -> Done
+# - Add columns to MembersList (e.g. Furnace Level, Power, Rank)
+# - Allow configuration of those additional columns via menubar
+# - add button to enter more embers in member list
+#NICE TO HAVE for 0.3:
+# - 
 # script dir will be used as initial for load/save
 script_dir=os.path.dirname(os.path.realpath(__file__))
 init_dir =os.path.join(script_dir,'hive')
 save_dir = os.path.join(init_dir,'save')
 
 class Member():
-    def __init__(self, number=0, name='', coords=' --- ',power=0,status='', 
+    def __init__(self, number=0, name='', coords='[XXX, YYY] ',power="100.0M",level='FC1',rank='R1',status='', 
                  city_id=None, widget = None, coord_widget = None, canvas = None):
         self.number=number
-        self.name = name
+        self.name = tk.StringVar(value=name)
         self.coords = coords
-        self.power = power
+        self.power = tk.StringVar(value=power)
+        self.level = tk.StringVar(value=level)
+        self.rank = tk.StringVar(value=rank)
         self.status = status
         self.city_id = city_id
         self.widget = widget
@@ -95,107 +109,6 @@ class Member():
             except AttributeError:
                 pass
 
-
-class Block():
-    multiplier = 5
-    def __init__(self,size=0, area=0,coords=[0,0], color='red',tag = '', id=dict(), floor_color=used_colors["floor"], text_color =used_colors["floor"]):
-        #create a block(building, floor, etc)
-        # with:
-        # size in grid coordinates
-        # area size of its floor (for FLag and HQ)
-        # color of the block on the canvas
-        # multiplier: Pixel per grid point
-        self.size = size
-        self.area = area
-        self.coords = coords
-        self.color = color
-        self.tag=tag
-        self.id = dict()
-        self.id.update(id)
-        self.floor_color = floor_color
-        self.text_color = text_color
-    
-    def paint(self,canvas,coords,grid=True):
-        self.canvas = canvas
-        if grid:
-            self.coords = self.convGrid2Coord(coords)
-        else:
-            self.coords=coords
-
-        building = canvas.create_rectangle(self.coords[0]-self.size*self.multiplier,self.coords[1]-self.size*self.multiplier,
-                                self.coords[0]+self.size*self.multiplier, self.coords[1]+self.size*self.multiplier, 
-                                fill=self.color,tags=('building',self.tag))
-        self.id.update({'building' : building})
-        canvas.buildings.append(self)
-        #if building provides Area, add it here
-        if self.area > 0:
-            floor=canvas.create_rectangle(self.coords[0]-self.area*self.multiplier,self.coords[1]-self.area*self.multiplier,
-                        self.coords[0]+self.area*self.multiplier, self.coords[1]+self.area*self.multiplier, 
-                        stipple="gray12", fill=self.floor_color,outline ='',tags=('floor',self.tag))
-            #floor is lowest unless in 'Show Floor' mode
-            if not canvas.top_floor:
-                canvas.tag_lower(floor)
-            else:
-                canvas.tag_raise(floor)
-            self.id.update({'floor' : floor})
-            return [building, floor]        
-        return building
-    
-    def convGrid2Coord(self,grid_coords):
-        #convert Grid coordinates into pixel coordinates on canvas
-        #Attention: for CANVAS coordinates POSITIVE Y is DOWN
-        #but for GRID coordinates POSITIVE Y is UP!
-        coords = [self.canvas.origin[0]+grid_coords[0]*self.multiplier*2, self.canvas.origin[1] - grid_coords[1]*self.multiplier*2]
-        return coords
-    
-    def box(self,center,grid = False,floor=False):
-        #gives upper left and lower right coordinates from center in 
-        # canvas coordinates (grid = False, default) or
-        # grid coordinates (grid = True)
-        # uses size for buildings and area for floor
-        if floor:
-            size=self.area
-        else:
-            size=self.size
-        if grid:
-            bbox =[center[0]-size/2, center[1]+size/2, 
-                   center[0]+size/2, center[1]-size/2]
-        else:
-            bbox =[center[0]-size*self.multiplier, center[1]-size*self.multiplier, 
-                   center[0]+size*self.multiplier, center[1]+size*self.multiplier]
-        return bbox
-
-class Flag(Block):
-     def __init__(self,size=1, area=7,coords=[0,0],color=used_colors["flag"],tag = 'Flag',floor_id=0):
-          super().__init__(size, area,coords,color,tag)
-          self.floor_id = floor_id  
-
-class City(Block):
-     def __init__(self,size=2, area=0,coords=[0,0],color=used_colors["city"], tag='City'):
-          super().__init__(size, area,coords,color,tag) 
-          #self.assign = assign     
-
-class HQ(Block):
-     def __init__(self,size=3, area=15,coords=[0,0],color=used_colors["hq"], tag='HQ',floor_id=0):
-          super().__init__(size, area,coords,color,tag)
-          self.floor_id = floor_id   
-        
-class Trap(Block):
-     def __init__(self,size=3, area=0,coords=[0,0],color=used_colors["trap"], tag='Trap'):
-          super().__init__(size, area,coords,color,tag)       
-
-class Rock(Block):
-     def __init__(self,size=2, area=0,coords=[0,0],color=used_colors["rock"], tag='Rock'):
-          super().__init__(size, area,coords,color,tag)  
-
-class Castle(Block):
-    def __init__(self,size=6, area=28,coords=[0,0],color=used_colors["castle"], tag='Castle'):
-          super().__init__(size, area,coords,color,tag) 
-
-class Tower(Block):
-    def __init__(self,size=2, area=0,coords=[0,0],color=used_colors["castle"], tag='Tower'):
-          super().__init__(size, area,coords,color,tag) 
-
 class MembersList(tk.Toplevel):
     # Defines the window with the list of member names
     def __init__(self,lines,geometry='100x100+100+100'):
@@ -203,44 +116,133 @@ class MembersList(tk.Toplevel):
         self.style = self.master.style
         self.lines = lines
         self.members=[]
+        self.column_mode = [tk.IntVar(value=1), tk.IntVar(value=1), tk.IntVar(value=1)]
+
         #build member list and show in seperate Window
         self.title("Alliance Members List")
         self.geometry(geometry)
         #add border
         self.config(bd=3, bg=used_colors['bg'])
 
-        upper_frame = ttk.Frame(self)
-        #upper_frame.pack()
-        upper_frame.pack(fill='x')
-        update_c_button=ttk.Button(upper_frame, text="Update", style='TButton', 
-                                   command=self.updateCoordinates)
-        update_c_button.grid(row=0,column=0, padx=(7, 7), pady=(7, 7), sticky='w')
-        close_button=ttk.Button(upper_frame, text="Close", style='TButton', command=self.remove)
-        #close_button.pack()
-        close_button.grid(row=0,column=1, padx=(7, 7), pady=(7, 7), sticky='e')
-
-        upper_frame.columnconfigure(0, weight=0)
-        upper_frame.columnconfigure(1, weight=1)
+        #canvas for List header
+        self.head_canvas = tk.Canvas(self,bg=used_colors['bg'],bd = 0, height=20, highlightthickness=0)
+        self.head_canvas.pack(fill='x')
 
         # frame with actual list
         # the data is in the frame inside the scroll frame
         self.scroll_frame = VerticalScrolledFrame(self)
         self.scroll_frame.pack(fill='both', expand=True)
         self.new_frame = self.scroll_frame.interior
-        self.new_frame.columnconfigure(1, weight=1)
+        #self.new_frame.columnconfigure(1, weight=1)
 
+        #self.head_canvas.columnconfigure(1, weight=1)
         #add members, to be filled in addMember
         for line in lines:
-            self.members.append(Member(name=line.strip()))
+            mem_attr = line.split(';')
+            #if power, level, and rank are there, use them, otherwise default them
+            if len(mem_attr) == 1:
+                self.members.append(Member(name=mem_attr[0].strip()))
+            elif len(mem_attr) == 2:
+                self.members.append(Member(name=mem_attr[0].strip(),power=mem_attr[1].strip()))
+            elif len(mem_attr) == 3:
+                self.members.append(Member(name=mem_attr[0].strip(),power=mem_attr[1].strip(),level=mem_attr[2].strip()))
+            elif len(mem_attr) == 4:
+                self.members.append(Member(name=mem_attr[0].strip(),power=mem_attr[1].strip(),
+                                    level=mem_attr[2].strip(),rank=mem_attr[3]))       
+            else:
+                if mem_attr[2].strip().isdecimal():
+                    mem_attr[2] = "FC"+mem_attr[2].strip()
+
+                self.members.append(Member(name=mem_attr[0].strip(),power=mem_attr[1].strip(),
+                                    level=mem_attr[2].strip(),rank=mem_attr[3]))
         self.setup()
+        self.update()
+        #adjust header
+        self.generateHeader()
+
+        # add (empty) menubar
+        self.mB = tk.Menu(self)
+        self['menu'] = self.mB
+        self.configMenu = tk.Menu(self.mB, tearoff=0)
+        #self.configMenu.add_command(label='Close',command=self.remove)
+        #set all columns to on by default:
+        self.power_mode = tk.IntVar(value=1)
+        self.lvl_mode = tk.IntVar(value=1)
+        self.rank_mode = tk.IntVar(value=1)
+        self.configMenu.add_checkbutton(label='Show Power',  
+                                      variable=self.column_mode[0], command=self.columnMode)
+        self.configMenu.add_checkbutton(label='Show Level',  
+                                      variable=self.column_mode[1], command=self.columnMode)
+        self.configMenu.add_checkbutton(label='Show Rank',  
+                                      variable=self.column_mode[2], command=self.columnMode)
+        self.configMenu.add_separator()
+        #move from update button to menu action to update Coordinates
+        self.configMenu.add_command(label='Update Coordinates',accelerator='Ctrl+U',command=self.updateCoordinates)
+        self.bind("<Control-u>", self.updateCoordinates)
+        self.configMenu.add_separator()
+        
+        #also remove the close butten and make it a command
+        self.configMenu.add_command(label='Close Members List',accelerator='Alt+F4',command=self.destroy)
+        # Alt-F4 is already defined, no need to bind it again
+        self.mB.add_cascade(label='Config', menu=self.configMenu)
+        self.update()
 
     def setup(self):
         self.ml = []
         self.names = []
+        self.power = []
+        self.level = []
         #for i, member in enumerate(self.members):
         for member in self.members:
             self.addMember(member)
- 
+
+    def generateHeader(self):
+        #remove old header widgets if any
+        self.head_canvas.delete(*self.head_canvas.find_all())
+            
+        header_names = ['ID','Member Name','Coords']
+        additional_columns = ['Power','Lvl', 'Rank']
+        for i, add_col in enumerate(additional_columns):
+            #if column is activated add header name
+            if self.column_mode[i].get():
+                header_names.append(add_col) 
+        for i, head in enumerate(header_names):
+            bbox = self.new_frame.grid_bbox(i, 0)
+            #add the padding in x and center in y
+            #self.head_canvas.create_text(bbox[0]+0,bbox[1]+10,text=head,anchor=tk.W,fill = 'white')
+            self.head_canvas.create_text(bbox[0]+bbox[2]/2,bbox[1]+int(self.head_canvas.cget('height'))/2,text=head,anchor=tk.CENTER,fill = 'white')
+            self.head_canvas.addtag_withtag(head,i)
+        #correct width of window
+        need = bbox[0]+bbox[2]
+        #DEBUG
+        self.new_frame.config(width=need)
+        self.scroll_frame.config(width=need)
+        #only change width with colum number
+        #height is at least as high as main window
+        #16 px is the width of the scrollbar (needs to be added to make it look better)
+        self.minsize(need+16, self.master.winfo_height())
+        self.maxsize(need+16, 2048)
+        #ttk.Label(self, width=need).pack()
+
+    def columnMode(self):
+        #change columns and build list anew
+        #destroy old list and build it anew
+        self.scroll_frame.destroy()
+        #and remove 'current' status from current member
+        cur_member = find(lambda my: "current line;" in my.status, self.members)
+        if cur_member:
+            cur_member.status =cur_member.status.replace("current line;","")
+        # the data is in the frame inside the scroll frame
+        self.scroll_frame = VerticalScrolledFrame(self)
+        self.scroll_frame.pack(fill='both', expand=True)
+        self.new_frame = self.scroll_frame.interior
+        #self.new_frame.columnconfigure(1, weight=1)
+
+        self.setup()
+        self.update()
+        #add and remove columns that are selected
+        self.generateHeader()
+
     def selectName(self,event, var):
         #select the klicked name
         member_name = var.get()
@@ -250,7 +252,7 @@ class MembersList(tk.Toplevel):
             old_cl.changeState("!current")
             old_cl.setColor()
 
-        member = find(lambda my: my.name == member_name, self.members)
+        member = find(lambda my: my.name.get() == member_name, self.members)
         #add new current line
         if not member:
             member=Member(name=member_name)
@@ -261,7 +263,7 @@ class MembersList(tk.Toplevel):
         #if yes, warn window
         #if not, destroy
         if find(lambda my: "assigned;" in my.status, self.members):
-            response = self.master.warn_window('Closing Members List will remove current assignments!\nDo you want to proceed?', buttons = 2, b_text=['Yes','No'])
+            response = self.master.warnWindow('Closing Members List will remove current assignments!\nDo you want to proceed?', buttons = 2, b_text=['Yes','No'])
         else:
             self.destroy() 
             return
@@ -273,11 +275,20 @@ class MembersList(tk.Toplevel):
             cities.clear()
             self.destroy()  
 
-    def updateCoordinates(self):
+    def updateCoordinates(self, event= None):
         canvas = self.master.paint_canvas
         for member in self.members:
             if "assigned;" in member.status:
                 canvas.showMemberCoords(member)
+        self.updateHeader()
+
+    def updateHeader(self): 
+        self.update()
+        for id in self.head_canvas.find_all():
+            new_x = self.new_frame.grid_bbox(id-1, 0)[0]
+            own_x = self.head_canvas.bbox(id)[0]
+            self.head_canvas.move(id,new_x-own_x,0)
+        self.head_canvas.update()
 
     def removeCityAssignment(self, city, member_name):
         #remove the city from the assignment list and delete the number
@@ -289,12 +300,12 @@ class MembersList(tk.Toplevel):
         #remove tag
         canvas.dtag(city, member_name)
 
-        member = find(lambda my: my.name == member_name,self.members)
+        member = find(lambda my: my.name.get() == member_name,self.members)
         if member:
             member.changeState("!assigned")
             member.city_id = None
             member.coords = [0, 0]
-            member.coord_widget.config(text=' --- ')
+            member.coord_widget.config(text='[XXX, YYY]')
 
         #remove from list of assigned cities
         #TODO: this doesn_t work with iterations
@@ -307,27 +318,76 @@ class MembersList(tk.Toplevel):
             #remove whitespaces and CR at start and end of name
             new_member=new_member.strip()
             #Is member in list already?
-            member = find(lambda my: my.name == new_member, self.members)
+            member = find(lambda my: my.name.get() == new_member, self.members)
             if not member:
                 self.members.append(Member(name=new_member))
                 self.addMember(self.members[-1])
 
+    def checkWidget(self,member, *argv):
+        #a change in the name has also to be traced to assigned cities, if any!
+        canvas = self.master.paint_canvas
+        new_name = member.name.get()
+        try:
+            city = canvas.cities[member.save_name]
+            canvas.cities.pop(member.save_name)
+            canvas.cities.update({new_name: city})
+            #also update the tag of the city in question
+            canvas.addtag_withtag(new_name, city)
+            #and remove old tag
+            canvas.dtag(city, member.save_name)
+            canvas.assignments[city] = new_name
+            #and bind the keys
+            #bind to the city (not the text!)
+            canvas.tag_bind(new_name,'<Enter>',func=lambda event: showAssignment(event=event,canvas = canvas))
+            canvas.tag_bind(new_name,'<Leave>',func=lambda event: showAssignment(event=event,canvas = canvas))
+        #exception if Member was not yet assigned
+        except KeyError:
+            pass
+        
+        #DEBUG
+        print(canvas.assignments[city],member.save_name,member.name.get())
+        #END DEBUG
+        #update old name
+        member.save_name = member.name.get()
+        return True
+
     def addMember(self, member):
         #add new member to MembersList
-        self.names.append(tk.StringVar(master =self.new_frame, value=member.name))
+        self.names.append(member.name)        
         new_idx = len(self.names)
+        member.save_name = member.name.get()
+        member.name.trace_add('write', lambda *pargs: self.checkWidget(member, *pargs)) #DEBUG
         member.number= new_idx
         ml_number = ttk.Label(self.new_frame,style='TLabel',text=str(new_idx))
         ml_number.grid(row=new_idx,column=0, sticky='news')
-        ml_entry = ttk.Entry(self.new_frame,style='TEntry',textvar=self.names[new_idx-1], width=20)
+        #Name entry: How to allow selection for assignment after name change?
+        #checkCommand= (self.register(self.checkWidget), '%s','%P','%W')
+        
+        ml_entry = ttk.Entry(self.new_frame,style='TEntry',textvar=member.name, width=20)
+                             #validate='key', validatecommand=checkCommand)
+
         ml_entry.grid(row=new_idx,column=1, sticky='news')
         member.widget = ml_entry
         ml_coords = ttk.Label(self.new_frame,style='TLabel',text=member.coords)
         ml_coords.grid(row=new_idx,column=2, padx= 7, sticky='news')
         member.coord_widget = ml_coords
-        
-        ml_entry.bind('<ButtonRelease-1>', func=lambda event,var = self.names[new_idx-1]: self.selectName(event,var))
-        self.ml.append((ml_number,ml_entry, ml_coords ))
+        # Add new columns here if neccessary
+        new_columns = [member.power, member.level,member.rank]
+        col_width = [6,4,3]
+        ml_extra= ['' for a in range(len(new_columns))]
+        #also specify number of columns that are fixed
+        num_fixed_col = 3
+        #if column is set, make Entry, otherwise skip
+        i=0
+        for c, col_conf in enumerate(self.column_mode):
+            if col_conf.get() == 1:
+                ml_extra[i]=ttk.Entry(self.new_frame,style='TEntry',textvar=new_columns[c],
+                           width=col_width[c])
+                ml_extra[i].grid(row=new_idx,column=num_fixed_col+i, padx= 2, sticky='news')
+                i+=1
+        ml_entry.bind('<Button-1>', func=lambda event,var = member.name: self.selectName(event,var))
+        #ml_entry.bind('<ButtonRelease-1>', func=lambda event,var = self.names[new_idx-1]: self.selectName(event,var))
+        self.ml.append((ml_number,ml_entry, ml_coords, *ml_extra))
 
         #change state of old 'current' member
         cur_member = find(lambda my: "current line;" in my.status, self.members)
@@ -341,442 +401,11 @@ class MembersList(tk.Toplevel):
         for memb in self.members:
             memb.changeState("!current")
 
-class IsoCanvas(tk.Canvas):
-    def __init__(self, master, width=0, height=0, bg='white', print_coords = True):
-        super().__init__(master=master,width=width, height=height, bg=bg)
-
-    def findAssignee(self,id):
-        #find the name of the member assigned to the id of the canvas object
-        ret_val = list(set(self.gettags(id)).intersection(set(self.cities.keys())))
-        #if only one member (which should always be the case) return string, not list
-        if len(ret_val) == 0:
-            return None
-        elif len(ret_val) == 1:
-            ret_val = ret_val[0] 
-        return ret_val
-    
-#TODO: make PaintCanvas scrollable to allow for larger hives
-class PaintCanvas(IsoCanvas):
-    def __init__(self, master, width=0, height=0, bg='white', print_coords = True):
-        super().__init__(master=master,width=width, height=height, bg=bg)
-        self.master = master
-        self.width = width
-        self.height = height
-        self.origin = [0,0]
-
-        #setup button defaults
-        self.erase = False
-        self.print_coords = print_coords
-        self.block = None
-        self.top_floor= False
-        self.grid_on = False
-        self.assign_mode = False
-        self.zoom_on = False
-        self.buildings=list()
-        self.cities = dict()
-        self.showMember = None
-        self.grid_size = 6
-
-        #define mouse actions
-        #Button 1
-        self.bind('<Button-1>', selectBlock)
-        self.bind('<B1-Motion>', moveBlock)
-        self.bind('<ButtonRelease-1>', putBlock)
-        
-        #generel movement
-        self.bind('<Motion>', self.showCoords)
-
-        # for area select use button-3
-        self.bind('<Button-3>', self.startSelect)
-        self.bind('<B3-Motion>', self.showArea)
-        self.bind('<ButtonRelease-3>', self.areaSelect)
-
-        #self.makeGrid()
-
-    def setup(self):
-        # resize canvas if window resizes
-        self.master.bind("<Configure>", self.on_resize)
-        # set origin of grid and activate grid
-        self.block = None
-        self.buildings=list()
-        self.cities = dict()
-        self.showMember = None
-        self.origin = [self.winfo_width()/2, self.winfo_height()/2]
-        self.makeGrid(self.grid_size)
-
-    def zoom(self, zoom_factor = 2.0):
-        self.zoom_on = not self.zoom_on
-        origin = self.origin
-        if not self.zoom_on:
-            self.master.zoom_button.state(['!pressed'])
-            zoom_factor = 1/zoom_factor
-        else:
-            self.master.zoom_button.state(['pressed'])
-        #remove coordnates
-        self.delete(self.find_withtag('c_coords'))
-        self.delete(self.find_withtag('g_coords'))
-        #up- or downscale everything
-        self.scale('all',*origin,zoom_factor,zoom_factor)
-        Block.multiplier *= zoom_factor
-        #re-draw the grid
-        self.makeGrid(self.grid_size)
-
-    def on_resize(self,event):
-        # when the canvas resizes, Grid and Coordinates need to be updated
-        # end all blocks have to move
-        old_origin = self.origin
-        origin = [self.winfo_width()/2,self.winfo_height()/2]
-        
-        dx = origin[0] - old_origin[0]
-        dy = origin[1] - old_origin[1]
-
-        #if sized changed: 
-        if dx != 0 or dy != 0:
-            # draw new grid
-            self.makeGrid(self.grid_size)
-            #remove coordnates
-            self.delete(self.find_withtag('c_coords'))
-            self.delete(self.find_withtag('g_coords'))
-            #move blocks: buildings, assigned numbers, floors, origin
-            blocks_to_move = set(self.find_withtag('building'))
-            blocks_to_move.update(self.find_withtag('member'),
-                                  self.find_withtag('floor'),
-                                  self.find_withtag('origin'))
-            for block in blocks_to_move:
-                self.move(block,dx,dy)
-
-    def convCoord2Grid(self,coords,block=None,trap_coords=None):
-        # size of block, if block is
-        #convert normal coordinates into grid positions and lock into grid
-        #Attention: for CANVAS coordinates POSITIVE Y is DOWN
-        #but for GRID coordinates POSITIVE Y is UP!   
-
-        #if a Block() is provided, use its size and multiplier
-        if isinstance(block,Block):
-            size=block.size 
-            multiplier=block.multiplier
-        #otherwise use default
-        else:
-            size=0
-            multiplier=Block.multiplier
-    
-        grid_coords = [0,0]
-
-        #objects of even size need to shift by 0.5 grid points as origin is at a half-point
-        grid_shift = (size+1)%2*0.5
-        #move to closest grid line, not only to the right...
-        #Y coord is negative, see above
-        dummy_c = [(coords[0]-self.origin[0])/(multiplier*2), (-coords[1]+self.origin[1])/(multiplier*2)]
-        round_c = [round((coords[0]-self.origin[0])/(multiplier*2)), round((-coords[1]+self.origin[1])/(multiplier*2))]
-        for i in range(2):
-            if dummy_c[i] > round_c[i]:
-                grid_coords[i] = round_c[i] + grid_shift
-            else:
-                grid_coords[i] = round_c[i] - grid_shift
-
-        #Add Trap Coordinates
-        if trap_coords is not None:
-            #Debug
-            grid_coords = listadd(grid_coords,trap_coords)
-        return grid_coords
-
-    def showCoords(self, event):
-        #show canvas coordinates in lower right corner
-        # and grid coordinates in lower left corner
-        trap_coord = None
-        self.master.trap_entry.state(['disabled'])
-        if self.print_coords:
-            # if Trap exists, activate Coordinate field
-            if self.find_withtag('Trap'):
-                self.master.trap_entry.state(['!disabled'])
-                trap_coord = self.master.trap_c
-            text='{}, {}'.format(str(event.x), str(event.y))
-            # convert canvas coordinates to grid coordinates
-            delta=Block.multiplier
-            grid = [event.x-delta, event.y+delta]
-            grid = self.convCoord2Grid(grid,trap_coords=trap_coord) 
- 
-            text_grid = 'Grid: {}, {}'.format(str(grid[0]), str(grid[1]))
-            c_coords = self.find_withtag('c_coords')
-            g_coords = self.find_withtag('g_coords')
-            if c_coords:
-                self.itemconfigure(c_coords, text=text)
-            else:
-                self.text_coord = self.create_text(self.winfo_width()-10, self.winfo_height()-10, 
-                                                                anchor=tk.SE, text=text, tag=('c_coords','coords'))
-                self.itemconfig(self.text_coord,font=('Helvetica', 15))
-            
-            if g_coords:
-                self.itemconfigure(g_coords, text=text_grid)
-            else:
-                self.text_grid_coord = self.create_text(10, self.winfo_height()-10, anchor=tk.SW, text=text_grid,tag=('g_coords','coords'))
-                self.itemconfig(self.text_grid_coord,font=('Helvetica', 15))
-    
-    def removeCoords(self):
-            #if coords are shown, delete them
-            try:
-                self.delete(self.text_coord)
-                self.delete(self.text_grid_coord)
-            except AttributeError:
-                pass
-    
-    def removeBuilding(self, building):
-        #remove building from the canvas and the building list     
-        block = self.getBuildingFromId(building) 
-        #is there a building with this ID?  
-        if block is not None:
-            self.buildings.remove(block)
-            self.delete(*block.id.values())
-            return
-        #if it's not in the building list, just delete it from the canvas
-        self.delete(building)
-    
-    def moveBuilding(self, building, center, grid=False, rel=False):
-        #move the selected building with all it:s appendices
-        if isinstance(building,Block):
-            #convert to canvas coords if grid=True
-            if grid:
-                center=building.convGrid2Coord(center)
-            #if it's relative coordinates, use Canvas.move, otherwise use Canvas.coords
-            
-            for id in building.id.values():
-                if rel:
-                    self.move(id, *center)
-                else: 
-                    floor = 'floor' in self.gettags(id)
-                    box = building.box(center,floor=floor)
-                    # the assigment indicator (canvas text) has only 2 coordinates, not 4
-                    if 'member' in self.gettags(id):
-                        box =[box[0]+2,box[1]+2]
-                    self.coords(id,*box)
-            #put assignment indicators on top
-            self.tag_raise('member','all')
-            #but still have the floor on top if so indicated
-            if self.top_floor:
-                self.tag_raise('floor','all')
-        else:
-            if rel:
-                self.move(building, *center)
-            else:
-                self.coords(building,*Block().box(center))
-                    
-    def startSelect(self, event):
-        #define upper left corner of area
-        self.area = [event.x, event.y , 0, 0]
-        #create rectangle w size 1
-        self.area_box = self.create_rectangle(event.x, event.y ,event.x+1, event.y+1 )
-
-    def showArea(self, event):
-        #define lower right corner, 
-        self.area[2:4] = [event.x, event.y]
-        #then change size of the rectangle
-        self.coords(self.area_box,*self.area)
-
-    def areaSelect(self, event):
-        #find all blocks inside the area
-        #check if selection was actually done by self.showArea and not just right click
-        #so release event has to be different to click event (otherwise size = 0)
-        if self.area[0:2] != [event.x, event.y]:
-            all_stuff = set(self.find_enclosed(*self.area))
-            selected = set(self.find_withtag('building'))
-            selected.union(set(self.find_withtag('floor')))
-            selected_stuff = all_stuff.intersection(selected)
-            for select in selected_stuff:
-                self.itemconfig(select,stipple='gray12')
-                self.addtag('selected','withtag', select)
-            #then release the selection frame
-            self.delete(self.area_box)    
-
-    def drawDefault(self):
-        #draw HQ, Trap and Grid
-        #clear canvas before drawing
-        self.delete('all')
-        self.master.resetTrapCoord()
-        self.buildings = []
-        #but add the grid
-        self.makeGrid(self.grid_size)
-        #remove assignments from member list (if any)
-        try:
-            #remove active player tag
-            for member in self.MembersList.members:
-                member.changeState("!assigned")
-            #self.master.MembersList.members_list.tag_remove('assigned','1.0','end')
-        except AttributeError:
-            pass
-        #put HQ and Trap
-        HeadQuarter = HQ(coords=[-11, 5]) 
-        NewTrap = Trap(coords = [0,0])
-
-        #Set Trap at origin
-        NewTrap.paint(self,NewTrap.coords)
-        #put origin marker
-        self.originMarker()        
-        #then HQ
-        HeadQuarter.paint(self,HeadQuarter.coords)
-        #floor is under buildings
-        self.tag_lower('floor','building')    
-
-    def originMarker(self):
-        #draw the origin
-        om = self.create_rectangle(self.winfo_width()/2-1,self.winfo_height()/2-1,self.winfo_width()/2+1,self.winfo_height()/2+1)
-        self.addtag_withtag('origin',om)
-
-    def makeGrid(self,grid_space = 6):
-        #draw grid on canvas
-        #first remove existing grid if any
-        self.removeGrid()
-        # grid_space: distance betwen grid lines in grid coordinates
-        self.origin = [self.winfo_width()/2,self.winfo_height()/2]
-        self.grid_on = True
-        multiplier = Block().multiplier*2
-        pix_space = grid_space*multiplier
-        # the origin is considered a 'half' grid point, so ensure that the lines reflect that
-        width = self.winfo_width()
-        height = self.winfo_height()
-        # initialize grid line:
-        grid_pos = grid_space/2
-        # if grid_space is an even number, add a half grid point to match full gridpoint roster1
-        grid_pos += (grid_space+1)%2*0.5
-        pix_pos = grid_pos*multiplier
-        while pix_pos < max(width, height):
-            #horizontal (plus and minus of origin)
-            self.create_line(0,self.origin[1]+pix_pos,width,self.origin[1]+pix_pos,dash=(1,5),tag='grid')
-            self.create_line(0,self.origin[1]-pix_pos,width,self.origin[1]-pix_pos,dash=(1,5),tag='grid')
-            #vertical
-            self.create_line(self.origin[0]+pix_pos,0,self.origin[0]+pix_pos,height,dash=(1,5),tag='grid')
-            self.create_line(self.origin[0]-pix_pos,0,self.origin[0]-pix_pos,height,dash=(1,5),tag='grid')
-            pix_pos += pix_space
-        
-        #create coordinate axes:
-        self.create_line(0,self.origin[1],width,self.origin[1],dash=(7,1,1,1),tag='grid')
-        self.create_line(self.origin[0],0,self.origin[0],height,dash=(1,1,1,1),tag='grid')
-        
-        #at the end put grid under buildings but above floor (only relevant for on/off)
-        if self.find_withtag('building'):
-            self.tag_lower('grid','building')
-
-    def removeGrid(self):
-        # delete grid
-        grid = self.find_withtag('grid')
-        self.delete(*grid)
-        self.grid_on = False
-
-    def changeGridSize(self,new_grid_size = None ):
-        #This is just a test function for now
-        #TODO: implement grid size selection on MainWindow level!
-        #use default if nothing is send
-        if new_grid_size == None:
-            new_grid_size = self.grid_size 
-        if new_grid_size == 6:
-            self.grid_size = 2 
-        else:
-            self.grid_size = 6
-        #then remove old grid
-        self.removeGrid()
-        #and paint the new one
-        self.makeGrid(grid_space=self.grid_size)
-
-
-    def getBuildingFromId(self,building_id):
-        #get the building in the self.buildings list from the id
-
-        for building in self.buildings:
-            #if building id in id values
-            if building.id['building'] == building_id:
-                return building
-        return None
-    
-    def assignMember(self,ML,city):
-        #assign selected member to building
-        member = find(lambda my: "current line;" in my.status, ML.members)
-        #if there is a current line
-        if not member:
-            return
-        member_num = member.number
-        member_name = member.name
-        
-        self.addtag_withtag('assigned',city)
-        #remove old assignement as one member can have only one city
-        if member_name in self.cities.keys():
-            old_city = self.cities[member_name]
-            ML.removeCityAssignment(old_city,member_name)
-            self.cities.pop(member_name)
-        #and remove previous owner of city (if any)
-        if city in self.cities.values():
-            old_owner = [k for k,v in self.cities.items() if v == city][0]
-            #remove old assignment
-            ML.removeCityAssignment(city,old_owner)
-            self.cities.pop(old_owner)
-
-        self.cities.update({member_name : city})
-        #also tag the city in question
-        self.addtag_withtag(member_name, city)
-        # and the other way around
-        self.assigments = { id : city for city, id in self.cities.items()}
-        member.city_id = city
-        member.canvas = self
-
-        # provide coordinates if trap coordinates are activated
-        if self.master.trap_c is not None:
-            self.showMemberCoords(member)
-
-        #bind to the city (not the text!)
-        self.tag_bind(member_name,'<Enter>',func=lambda event: showAssignment(event=event,canvas = self))
-        self.tag_bind(member_name,'<Leave>',func=lambda event: showAssignment(event=event,canvas = self))
-
-        #then show the number on the city
-        x0, y0, *_ = self.coords(city)
-        text_id=self.create_text(x0+2, y0+2, anchor=tk.NW, text=str(member_num), fill='black',activefill='white',tag=('member'))
-        # get the building object
-        block = self.getBuildingFromId(city)
-        if block is not None:
-            block.id.update({'text':text_id})
-        #set textcolor according to block setting
-        self.itemconfig(text_id,fill=used_colors["assign"])
-        self.tag_lower('building','member')
-        #also mark it in member list and remove the current selection;:
-        member.changeState("new assigned,!current")
-        #autoselect next unassigned member
-        # remark: index produces an error when member does not exist in the list, but as it was just extracted from the list, it has to be there anyway
-        cur_mem = ML.members.index(member)
-        #check next member for assognment
-        self.selectNextMember(ML,cur_mem)
-  
-    def selectNextMember(self,MList,cur_member,found_member = False):
-        #select the next unassigned member in MemberList after cur_member     
-        for next_member in MList.members[cur_member+1:]:
-            if "assigned" not in next_member.status:
-                next_member.changeState("new current")
-                return True
-        #no member found after cur_member? Just start again at the beginning
-        if found_member is False:
-            found_member = self.selectNextMember(MList,-1,True)
-            #we have to return even if the complete list is already assigned
-            #TODO: give warning that all members have been assigned if found_member is still False
-            return True
-
-    def showMemberCoords(self, member):
-        #Calculate the city coordinate in the trap grid and display it
-        #check if trap coordinates are active
-        if self.master.trap_c is not None:
-            cc = self.coords(member.city_id)
-            cc = [cc[0],cc[-1]]
-            cc_grid = [int(dummy) for dummy in self.convCoord2Grid(cc, trap_coords=self.master.trap_c)]
-            #change to new class
-            member.coords = cc_grid
-            member.coord_widget.config(text=str(member.coords))
-        else:
-            member.coord_widget.config(text=' --- ')
-
-    def findAssignee(self,id):
-        #find the name of the member assigned to the id of the canvas object
-        ret_val = list(set(self.gettags(id)).intersection(set(self.cities.keys())))
-        #if only one member (which should always be the case) return string, not list
-        if len(ret_val) == 0:
-            return None
-        elif len(ret_val) == 1:
-            ret_val = ret_val[0] 
-        return ret_val
+    def getLines(self):
+        member_lines = []
+        for member in self.members:
+            member_lines.append(';'.join([member.name.get(),member.power.get(),member.level.get(),member.rank.get()]))
+        return member_lines
 
 class MainWindow(tk.Tk):
     def __init__(self):
@@ -795,45 +424,55 @@ class MainWindow(tk.Tk):
         #get style
         self.style = initStyle(self)
 
+        #add list of buildings
+        #self.buildings = set(['HQ','City','Flag','Trap','Rock','Castle','Tower'])
+        self.buildings = ['City','Flag','HQ','Rock','Trap','Castle','Tower']
+
         #add border
         self.config(bd=3, bg=used_colors['bg'])
 
-        #Frame
+        #Frame for Buttons
         bt_frame=ttk.Frame(self, width = wwidth, height = 65)
         bt_frame.grid(row=0,column=0,sticky='news')
         bt_frame.columnconfigure(0,weight=1)
         
-        #setup Canvas
-        self.paint_canvas = PaintCanvas(self, width=wwidth, height = wheight, bg='white')
-
+        #Frame for Canvas
+        self.cs_frame= ScrolledFrame(self, width = wwidth, height = wheight)
+        self.int_frame = self.cs_frame.interior
+        self.int_frame.columnconfigure(0, weight=1)
+        self.int_frame.rowconfigure(0, weight=1)
+        self.upper_canvas = self.int_frame.master
+        self.paint_canvas = PaintCanvas(self.int_frame, width=wwidth*2, height = wheight*2, bg=used_colors["canvas_bg"])
+        
         #paint canvas
-        self.paint_canvas.grid(row=1,column=0,sticky='news')
+        self.cs_frame.grid(row=1,column=0,sticky='news')
+        self.paint_canvas.pack(side=tk.LEFT,fill='both', expand=True)
+        
         self.rowconfigure(1, weight=1)
         self.columnconfigure(0, weight=1)
+        self.upper_canvas.configure(width = wwidth, height = wheight)
+        #calculate scrollregion
+        self.update()
+        size = (int(self.paint_canvas.cget('width')), int(self.paint_canvas.cget('height')))
+        self.cs_frame.scrollArea(size)
 
         #first column:
         fc = 1
         self.bt_frame = bt_frame
-        #load/save/etc
-        self.save_button = ttk.Button(bt_frame, text='Save', command=self.saveLayout)
-        self.save_button.grid(row = 0, column=fc+0, sticky='news',padx=fpadx, pady=fpady)
-
-        self.load_button = ttk.Button(bt_frame, text='Load', command=self.loadLayout)
-        self.load_button.grid(row = 0, column=fc+1, sticky='news',padx=fpadx, pady=fpady)
-
+        
         #add a bit of space in between
         self.dummy= ttk.Label(bt_frame)
         self.dummy.grid(row = 0, column=fc+7, sticky='news',padx=30, pady=fpady)
         #self.columnconfigure(fc+7, weight=1)
 
         self.isoview_button = ttk.Button(bt_frame, text='Isometric', style='TButton', command=self.isometricView)
-        self.isoview_button.grid(row = 0, column=fc+8, sticky='news',padx=fpadx, pady=fpady)
+        self.isoview_button.grid(row = 0, column=fc+9, sticky='news',padx=fpadx, pady=fpady)
         
         self.default_button = ttk.Button(bt_frame, text='Default', style='TButton', command=self.default)
-        self.default_button.grid(row = 0, column=fc+9, sticky='nes',padx=fpadx, pady=fpady)
+        self.default_button.grid(row = 0, column=fc+10, sticky='nes',padx=fpadx, pady=fpady)
 
-        self.members_button = ttk.Button(bt_frame, text='Member List', command=self.fileMembersList)
-        self.members_button.grid(row = 0, column=fc+10, sticky='news',padx=fpadx, pady=fpady)
+        #self.members_button = ttk.Button(bt_frame, text='Member List', command=self.fileMembersList)
+        #self.members_button.grid(row = 0, column=fc+10, sticky='news',padx=fpadx, pady=fpady)
 
         self.assign_button = ttk.Button(bt_frame, text='Assign',style='Assign.TButton', command=self.assignMode)
         self.assign_button.grid(row = 0, column=fc+11, sticky='news',padx=fpadx, pady=fpady)
@@ -901,9 +540,13 @@ class MainWindow(tk.Tk):
         bt_frame.grid_propagate(False)
 
         self.iconbitmap(default = os.path.join(init_dir,'icon.ico'))
+        
+        #add menues
+        self.menues()
 
         #draw default after starting mainloop
-        self.after(100,self.setup)
+        self.after(100,self.setup())
+        
 
     #function for Trap Coordinate Entry field
     def validateCoord(self, new_text):
@@ -932,13 +575,143 @@ class MainWindow(tk.Tk):
                     return True
                 except ValueError:
                     return True
-                
+
+    def menues(self):
+        top = self.winfo_toplevel()
+        self.menuBar = tk.Menu(top)
+        top['menu'] = self.menuBar
+        # Menu for File Options
+        self.fileMenu = tk.Menu(self.menuBar, tearoff=0)
+        self.fileMenu.add_command(label='Save Hive',accelerator='Ctrl+S',command=self.saveLayout)
+        self.bind("<Control-s>", self.saveLayout)
+        self.fileMenu.add_command(label='Load Hive',accelerator='Ctrl+L',command=self.loadLayout)
+        self.bind("<Control-l>", self.loadLayout)
+        self.fileMenu.add_separator()
+        self.fileMenu.add_command(label='Save MembersList',accelerator='Ctrl+M', command=self.saveMembersList)
+        self.bind("<Control-m>", self.saveMembersList)
+        self.fileMenu.add_command(label='Load MembersList',accelerator='Ctrl+O', command=self.loadMembersList)
+        self.bind("<Control-o>", self.loadMembersList)
+        self.fileMenu.add_separator()
+        self.fileMenu.add_command(label='Save Color Theme',accelerator='Ctrl+T', command=self.saveColors)
+        self.bind("<Control-t>", self.saveColors)
+        self.fileMenu.add_command(label='Load Color Theme',accelerator='Ctrl+Shift+T', command=self.loadColors)
+        self.bind("<Control-T>", self.loadColors)
+        self.fileMenu.add_separator()
+        self.fileMenu.add_command(label='Quit',accelerator='Ctrl+Q', command=self.quitDialog)
+        self.bind("<Control-q>", self.quitDialog)
+        #Menu for Mode Options
+        self.modeMenu = tk.Menu(self.menuBar, tearoff=0)
+        self.modeMenu.add_checkbutton(label='Assign Mode', accelerator='Ctrl+A',  
+                                      variable=self.paint_canvas.assign_mode, command=self.assignMode)
+        self.bind("<Control-a>", self.assignMode)
+        self.modeMenu.add_checkbutton(label='Erase Mode', accelerator='Ctrl+E', 
+                                      variable=self.paint_canvas.erase, command=self.eraseBuilding)
+        self.bind("<Control-e>", self.eraseBuilding)
+
+        #Menu for Colors
+        self.colorMenu = tk.Menu(self.menuBar, tearoff=0)
+        self.BuildingColors = tk.Menu(self.colorMenu, tearoff=0)
+        for building in self.buildings:
+            #how to prevent that building at command call is always last in list ('Tower')        
+            self.BuildingColors.add_command(label=building, activebackground = used_colors[building.lower()], 
+                                        command = lambda b=building: self.getcolor(b))
+        #and one for the floor
+        self.BuildingColors.add_command(label='Floor', activebackground = used_colors['floor'], 
+                                        command = lambda: self.getcolor('floor'))
+        self.colorMenu.add_cascade(label='Buildings', menu=self.BuildingColors)
+        self.colorMenu.add_separator()
+        self.colorMenu.add_command(label='Background', activebackground = used_colors['canvas_bg'],
+                                   command = lambda: self.getcolor('canvas_bg'))
+        self.colorMenu.add_separator()
+        self.colorMenu.add_command(label='Default', command = lambda: self.getcolor('default'))
+        self.menuBar.add_cascade(label='File', menu=self.fileMenu)
+        self.menuBar.add_cascade(label='Modes', menu=self.modeMenu)
+        self.menuBar.add_cascade(label='Colors', menu=self.colorMenu)
+
+        self.update()
+
     def setup(self):
         #performed once mainloop is started
         self.paint_canvas.setup()
         self.active_button = None
-        self.buildings = set(['HQ','City','Flag','Trap','Rock','Castle','Tower'])
+        self.ML_width = 321
 
+    def quitDialog(self,event=None,):
+        #Ask for save hive or quit without saving
+        #but only if anything exists on the canvas
+        buildings_exist = max([len(self.paint_canvas.find_withtag(b)) for b in self.buildings ]) > 0
+        if not buildings_exist:
+            #without buildings, just quit
+            self.quit()
+            return
+        # if there is at least one building, ask for save
+        response = self.warnWindow('Save Hive before quitting??', buttons = 3, b_text=["Save Hive","Don't Save","Abort"])
+        if response =="abort":
+            return
+        elif response=="save hive":
+            self.saveLayout()
+        self.quit()
+
+    def getcolor(self, building):
+        #change colors of buildings and floor
+        global used_colors
+
+        def ChangeBuildingColor(building):
+            #find all buildings of the type
+            # and change to new color
+            new_paint=set(self.paint_canvas.find_withtag(building))
+            #only change the color of the building, not its floor
+            if building != 'floor':
+                exclude_floor = set(self.paint_canvas.find_withtag('floor'))
+                new_paint = new_paint.difference(exclude_floor)
+            for id in new_paint:
+                self.paint_canvas.itemconfig(id,fill=used_colors[building.lower()])
+            
+            #change button background
+            button = building.capitalize()
+            #HQ is special
+            if button == "Hq":
+                button = "HQ"
+            if button in ['City','Flag','Rock', 'HQ','Trap']:
+                self.style.map(button+'.Build.TButton', 
+                    background=[('active', used_colors[button.lower()]),
+                    ('!active','pressed', used_colors[button.lower()]), 
+                    ('!active','white')]) 
+        bblocks = deepcopy(self.buildings)
+        #TODO: Not nice, can we use used_colors.keys() instead?
+        bblocks.append('floor')
+        bblocks.append('canvas_bg')
+        #find position of canvas_bg in menu
+        bg_pos = findMenuIndex(menu = self.colorMenu, label = "Background")
+        if building.lower() in ['default', 'all']:
+            if building.lower() == 'default':
+                #reset all colors to default values
+                for key in used_colors_default.keys():
+                    used_colors[key] = used_colors_default[key]
+            # repaint everything
+            for block in bblocks:
+                if block == 'canvas_bg':
+                    self.paint_canvas.config(bg = used_colors[block])
+                    #keep the separator in mind!
+                    self.colorMenu.entryconfig(bg_pos, activebackground = used_colors[block.lower()])
+                else:
+                    ChangeBuildingColor(block)
+                    self.BuildingColors.entryconfig(bblocks.index(block), activebackground = used_colors[block.lower()])
+            return
+        #get the new color and put it in the color directory
+        new_color = askcolor(title='New Color')
+        used_colors[building.lower()] = new_color[1]
+
+        #if it's the background, don't change any buildings ;)
+        if building.lower() == 'canvas_bg':
+            self.paint_canvas.config(bg = new_color[1])
+            #keep the separator in mind!
+            self.colorMenu.entryconfig(bg_pos, activebackground = used_colors[building.lower()])
+            return
+        #change buildings of the changed type to the new color
+        self.BuildingColors.entryconfig(bblocks.index(building), activebackground = used_colors[building.lower()])
+        ChangeBuildingColor(building)
+        
     def resetTrapCoord(self):
         #TODO: make inactive even if trap is there
         self.trap_c = None 
@@ -975,7 +748,7 @@ class MainWindow(tk.Tk):
 
         top.mainloop()
 
-    def warn_window(self, warning,  buttons=1, b_text=['OK'], width=300, height=100):
+    def warnWindow(self, warning,  buttons=1, b_text=['OK'], width=300, height=100):
         # pop-up wwarning window
         # buttons: number of buttons in window
         # b_text: list of texts on button 
@@ -1025,7 +798,7 @@ class MainWindow(tk.Tk):
         #check if there are any buildings already on the canvas
         #if so warn about deletion
         if self.paint_canvas.find_withtag('building'):
-            response = self.warn_window('Default setting will delete current setup!\nDo you want to proceed?', buttons = 2, b_text=['Yes','No'])
+            response = self.warnWindow('Default setting will delete current setup!\nDo you want to proceed?', buttons = 2, b_text=['Yes','No'])
         else:
             response = 'yes'
         if response == 'yes':
@@ -1052,14 +825,14 @@ class MainWindow(tk.Tk):
 
     def printHQ(self):
         if self.paint_canvas.find_withtag('HQ'):
-            self.warn_window('Only one HQ allowed!\nHQ already exists')
+            self.warnWindow('Only one HQ allowed!\nHQ already exists')
         else:
             self.paint_canvas.block = HQ()
             self.activateButton(self.hq_button)
     
     def printTrap(self):
         if len(self.paint_canvas.find_withtag('Trap')) >= 2:
-            self.warn_window('Only two Traps allowed!\nTraps already exist!')
+            self.warnWindow('Only two Traps allowed!\nTraps already exist!')
         else:
             self.paint_canvas.block = Trap()
             self.activateButton(self.trap_button)
@@ -1099,25 +872,27 @@ class MainWindow(tk.Tk):
     def eraseBuilding(self):
         #toggle erase mode
         self.activateButton(self.erase_button) 
-        if not self.paint_canvas.erase:
-            self.paint_canvas.erase = True
+        if not self.paint_canvas.erase.get():
+            self.paint_canvas.erase.set(True)
             self.erase_button.state(['pressed'])
             self.paint_canvas.config(cursor='tcross')
+            #can't have assign and erase at the same time
+            self.assign_button.state(['!pressed'])
         else:
             self.resetErase()
     
     def resetErase(self):
-        self.paint_canvas.erase = False
+        self.paint_canvas.erase.set(False)
         self.erase_button.state(['!pressed'])
         self.paint_canvas.config(cursor='left_ptr')
     
-    def assignMode(self):
+    def assignMode(self,event=False):
         # check of the window with the member names exists
         try:
             self.MembersList.winfo_exists()
         #if not ask for load
         except AttributeError:
-            response = self.warn_window('No Member List available!\nLoad Member List?', buttons = 2, b_text=['Yes','No'])
+            response = self.warnWindow('No Member List available!\nLoad Member List?', buttons = 2, b_text=['Yes','No'])
             if response == 'yes':
                 self.loadMembersList()
             return
@@ -1127,14 +902,16 @@ class MainWindow(tk.Tk):
         # check if assign button is active
         #if self.assign_button.cget('relief') == 'sunken':
         if 'pressed' in self.assign_button.state():
-            self.paint_canvas.assign_mode = True
+            self.paint_canvas.assign_mode.set(True)
             #self.assign_button.config(bg = used_colors['current'])
             self.paint_canvas.config(cursor='man')
+            #can't have assign and erase at the same time
+            self.erase_button.state(['!pressed'])
         else:
             self.resetAssign()
 
     def resetAssign(self):
-            self.paint_canvas.assign_mode = False
+            self.paint_canvas.assign_mode.set(False)
             self.assign_button.state(['!pressed'])
             self.paint_canvas.config(cursor='left_ptr')
 
@@ -1206,7 +983,7 @@ class MainWindow(tk.Tk):
         obj_info_str = tag+': '+str(obj_info)+'\n'
         return obj_info_str.encode('utf8', errors='xmlcharrefreplace')
 
-    def saveLayout(self):
+    def saveLayout(self,event=False):
         #save buildings and assignments to file
         save_file = asksaveasfilename(title='Save Hive as:', defaultextension='.hb', initialdir=save_dir,filetypes=[('Hive Organizer file','*.hb')])
         self.assigments = { id : city for city, id in self.paint_canvas.cities.items()}
@@ -1220,7 +997,7 @@ class MainWindow(tk.Tk):
             building_info.append(self.buildingInfoEncoder(building))
         try:
             self.MembersList.winfo_exists()
-            member_list = self.MembersList.lines
+            member_list = self.MembersList.getLines()
             #strip whitespaces
             member_list = [member.strip() for member in member_list]
         except AttributeError:
@@ -1244,15 +1021,31 @@ class MainWindow(tk.Tk):
             if member_list is not None:
                 line = 'MemberList: '+'\n'.join(member_list)
                 file.write(line.encode('utf8', errors='xmlcharrefreplace'))
+            #save used colors if they differ from default
+            #Perhaps save even if default?
+            # TODO: Might also save everything as json
+            # But needs a legacy load function then
+            if used_colors is not used_colors_default:
+                col_file = save_file.split('.')[-2]+'.col'
+                with open(col_file, 'w') as fp:
+                    json.dump(used_colors, fp)
+
         #set zoom_status back to original setting
         if zoom_status:
             self.paint_canvas.zoom()
 
-    def loadLayout(self):
+    def loadLayout(self,event=False):
         #load previously saved layout
         lines=''
         ml_data = str()
         load_file = askopenfilename(title='Load Hive file:',initialdir=save_dir,defaultextension='.hb',filetypes=[('Hive Organizer file','*.hb'), ('All files','*.*')])   
+        #if a color file exists, load it before the rest
+        try:
+            col_file = load_file.split('.')[-2]+'.col'
+            self.loadColors(col_file=col_file)          
+        except FileNotFoundError:
+            #no associated color file, use current colors
+            pass
         if load_file:
             with open(load_file,'r',encoding='utf8') as file:
                 lines=file.readlines()
@@ -1281,12 +1074,12 @@ class MainWindow(tk.Tk):
                     if self.MembersList.winfo_exists():
                         self.updateMembersList(ml_data)
                     else:
-                        self.MembersList = MembersList(ml_data,geometry=("{}x{}+{}+{}".format(250, self.winfo_height(),self.winfo_x()+self.winfo_width(),self.winfo_y())))
+                        self.MembersList = MembersList(ml_data,geometry=("{}x{}+{}+{}".format(self.ML_width, self.winfo_height(),self.winfo_x()+self.winfo_width(),self.winfo_y())))
                         for member in self.MembersList.members:
                             member.changeState('!current')           
                 # there is no list
                 except AttributeError:
-                    self.MembersList = MembersList(ml_data,geometry=("{}x{}+{}+{}".format(250, self.winfo_height(),self.winfo_x()+self.winfo_width(),self.winfo_y())))
+                    self.MembersList = MembersList(ml_data,geometry=("{}x{}+{}+{}".format(self.ML_width, self.winfo_height(),self.winfo_x()+self.winfo_width(),self.winfo_y())))
                     #remove active player tag
                     self.MembersList.removeCurrent()
 
@@ -1294,8 +1087,33 @@ class MainWindow(tk.Tk):
             self.resetTrapCoord()
             for line in lines:
                 self.buildHive(line)
+            #update header of Memberslist if it is included in the save file
+            if ml_data:
+                self.MembersList.updateHeader()
             #put origin marker
             self.paint_canvas.originMarker()   
+
+    def saveColors(self,event=None,col_file=None):
+        #if it is called by hot_key, event is present and has to be adressed
+        #if no file name is provided, ask for one
+        if col_file is None:
+            col_file = asksaveasfilename(title='Save Color Theme:',initialdir=save_dir, defaultextension='.col', filetypes=[('Color Theme','*.col'), ('All files','*.*')]) 
+        try:
+            with open(col_file, 'w') as fp:
+                json.dump(used_colors, fp)
+        except FileNotFoundError:
+            self.warnWindow('Colors could not be saved!\nProvide correct Filename!')
+    
+    def loadColors(self, event=None, col_file=None):
+        #if it is called by hot_key, event is present and has to be adressed
+        #if no file name is provided, ask for one
+        if col_file is None:
+            col_file = askopenfilename(title='Load Color Theme:',initialdir=save_dir, defaultextension='.col', filetypes=[('Color Theme','*.col'), ('All files','*.*')]) 
+        with open(col_file, 'r') as fp:
+            used_colors_load = json.load(fp)
+        for key in used_colors_load.keys():
+            used_colors[key] = used_colors_load[key]
+        self.getcolor('all')
 
     def buildHive(self,line):
         #paint the Hive from the saved file
@@ -1341,7 +1159,7 @@ class MainWindow(tk.Tk):
                 coords=(listadd(list(map(float,ct)),self.paint_canvas.origin))
                 #then paint them:
                 build_new = eval(command)
-                build_new.paint(canvas, coords,grid=False)
+                build_new.paint(canvas, coords,grid=False,color=used_colors[block.lower()])
                 #for cities, add they have assigned members
                 if block == 'City':
                     for member in assignments:
@@ -1355,7 +1173,7 @@ class MainWindow(tk.Tk):
                             city=build_new.id['building']
 
                             #check if member exists in list
-                            member = find(lambda f: f.name == member_name, ML.members)
+                            member = find(lambda f: f.name.get() == member_name, ML.members)
                             if member:
                                 #remove active player tag
                                 self.MembersList.removeCurrent()
@@ -1370,21 +1188,7 @@ class MainWindow(tk.Tk):
         except AttributeError:
             pass
 
-    def fileMembersList(self):
-        #if MembersList already exists, save or load List
-        try:
-            ml_ok = self.MembersList.winfo_exists()
-            response = self.warn_window('Member List already active!\nSave current List or load new List?', 
-                                    buttons = 2, b_text=['Save','Load'])
-        except:
-            self.loadMembersList()
-            return     
-        if ml_ok and response == 'save':
-            self.saveMembersList()
-        else:
-            self.loadMembersList()
-
-    def loadMembersList(self):
+    def loadMembersList(self,event=False):
         load_file = askopenfilename(title='Load Member List:',initialdir=save_dir, defaultextension='.txt', filetypes=[('Member List text file','*.txt'), ('All files','*.*')]) 
         if load_file:
             with open(load_file,'r',encoding='utf8') as file:
@@ -1397,17 +1201,17 @@ class MainWindow(tk.Tk):
         #either there never was a members list or the window was closed:
         except AttributeError:
             pass
-        self.MembersList = MembersList(lines,geometry=("{}x{}+{}+{}".format(250, self.winfo_height(),self.winfo_x()+self.winfo_width(),self.winfo_y())))
+        self.MembersList = MembersList(lines,geometry=("{}x{}+{}+{}".format(self.ML_width, self.winfo_height(),self.winfo_x()+self.winfo_width(),self.winfo_y())))
 
-    def saveMembersList(self):
+    def saveMembersList(self,event=False):
         #save MemberslList to Textfile
         try:
             ml_ok = self.MembersList.winfo_exists()
-        except AttributeError:
+        except AttributeError or UnboundLocalError:
             return
         if ml_ok:  
             # get list of member names
-            member_list = self.MembersList.lines
+            member_list = self.MembersList.getLines()
             #strip whitespaces
             member_list = [member.strip() for member in member_list]          
             save_file = asksaveasfilename(title='Save Member List:',initialdir=save_dir, defaultextension='.txt', filetypes=[('Member List text file','*.txt'), ('All files','*.*')]) 
@@ -1419,13 +1223,20 @@ class MainWindow(tk.Tk):
 
     def updateMembersList(self,ml_data):
         #what to do if new List is loaded when list already exists
-        response = self.warn_window('Member List already active!\nOverwrite existing Member List,\n try to merge both lists or ignore list in File?', 
+        response = self.warnWindow('Member List already active!\nOverwrite existing Member List,\n try to merge both lists or ignore list in File?', 
                                     buttons = 3, b_text=['Overwrite','Merge','Ignore'])
         if response == 'overwrite':
+            #remove all old assignments
+            for member_name in self.paint_canvas.cities.keys():
+                old_city = self.paint_canvas.cities[member_name]
+                self.MembersList.removeCityAssignment(old_city,member_name)
+            #and set the list to 'empty'
+            self.paint_canvas.cities.clear()
+
             #remove existing list
             self.MembersList.destroy()
             #and generate new one
-            self.MembersList = MembersList(ml_data,geometry=("{}x{}+{}+{}".format(250, self.winfo_height(),self.winfo_x()+self.winfo_width(),self.winfo_y())))
+            self.MembersList = MembersList(ml_data,geometry=("{}x{}+{}+{}".format(self.ML_width, self.winfo_height(),self.winfo_x()+self.winfo_width(),self.winfo_y())))
             #remove active player tag
             for member in self.MembersList.members:
                 member.changeState('!current')
@@ -1451,7 +1262,31 @@ class MainWindow(tk.Tk):
         #upper_frame.pack()
         upper_frame.pack(fill='x')
         top.geometry("{}x{}+{}+{}".format(self.winfo_width(), self.winfo_height()+40,self.winfo_x()+20,self.winfo_y()+20))
-        top_canvas = IsoCanvas(top, width=self.winfo_width(), height=self.winfo_height(),bg='white')
+        
+        #Frame for Canvas
+        cs_frame= ScrolledFrame(top, width = top.winfo_width(), height = top.winfo_height())
+        int_frame = cs_frame.interior
+        int_frame.columnconfigure(0, weight=1)
+        int_frame.rowconfigure(0, weight=1)
+        upper_canvas = int_frame.master
+        
+        #isometric canvas
+        top_canvas = IsoCanvas(int_frame, width=self.paint_canvas.winfo_width(), 
+                               height=self.paint_canvas.winfo_height(),bg=used_colors["canvas_bg"])
+        
+        #cs_frame.grid(row=1,column=0,sticky='news')
+        cs_frame.pack(side=tk.LEFT,fill='both', expand=True)
+        top_canvas.pack(side=tk.LEFT,fill='both', expand=True)
+        
+        top.rowconfigure(1, weight=1)
+        top.columnconfigure(0, weight=1)
+        upper_canvas.configure(width = top.winfo_width(), height = top.winfo_height())
+        #calculate scrollregion
+        self.update()
+        size = (int(top_canvas.cget('width')), int(top_canvas.cget('height')))
+        cs_frame.scrollArea(size,2)       
+        #TODO: focus on the same area as PaintCanvas
+
         tmpng = ttk.Button(upper_frame, text="Make PNG", style='TButton', command=lambda: make_png(top_canvas,top))
         tmpng.grid(row=0,column=0, padx=(7, 7), pady=(7, 7), sticky='w')
         tcb = ttk.Button(upper_frame, text="Close", style='TButton', command=top.destroy)
@@ -1459,8 +1294,7 @@ class MainWindow(tk.Tk):
         top_canvas.pack()
         top_canvas.showMember = None
         top_canvas.cities = canvas.cities
-        cos_val = math.cos(math.radians(angle))
-        sin_val = math.sin(math.radians(angle))
+
         all_stuff = canvas.find_all()
 
         # go through every element and rotate it around the center
@@ -1492,12 +1326,9 @@ class MainWindow(tk.Tk):
 
             for i, c in enumerate(poly_coords):
                 # for each x/y set
-                #if i%2 == 0:
-                # x_rot = x*cos(angle) - y*sin(angle) + x_center
-                x_rot = c[0] * cos_val - c[1] * sin_val + canvas.origin[0]
-                # y_rot = x*sin(angle) + y*cos(angle) + y_center
-                y_rot = c[0] * sin_val + c[1] * cos_val + canvas.origin[1]
-                rot_coords[i] = [x_rot, y_rot]
+                #rotate the coordinates and add the center    
+                rot_coords[i] = listadd(rot(c, angle), canvas.origin)
+                #[x_rot, y_rot]
 
             # now put the rotated element on the new canvas
             if 'grid' in tags:
@@ -1525,329 +1356,9 @@ class MainWindow(tk.Tk):
                #move it a bit for better readability
                rot_coords = listadd(rot_coords[0],[5, -7])
                top_canvas.create_text(rot_coords, text_config) 
-              
-def getPoly(coords):
-    # provides polygon coordinates from rectangle coordinates
-    poly_coords = list()
-    # first top left, same as rectangle
-    poly_coords.append([coords[0],coords[1]])
-    # lower left
-    poly_coords.append([coords[0],coords[3]])
-    # lower right
-    poly_coords.append([coords[2],coords[3]])
-    # upper right
-    poly_coords.append([coords[2],coords[1]])
-    return poly_coords
-
-def showAssignment(event, canvas):
-    #show name of assigned member in a box
-    #get active object
-
-    current = canvas.find_withtag(tk.CURRENT)[0]
-    #get member from tag of event
-    member = canvas.findAssignee(current)
-    #list(set(canvas.gettags(current)).intersection(set(canvas.cities.keys())))[0] 
-
-    if event.type == tk.EventType['Enter']:
-        coords = canvas.coords(current)
-        #cur_cent = center(coords)
-        #city_size = City().size*City().multiplier*2
-        #coordinates for box relativ to city
-        # get minimal y-value, whether it's a rectangle or a polygon
-        coords[1] = min(list(map(lambda x:coords[x], range(1,len(coords),2))))
-        # the width of '0' is only a placeholder.
-        # it is made to fit the text below
-        coords_rec = [coords[0], coords[1]-20, coords[0], coords[1]]
-
-        # Create box for hover text
-        canvas.showMember = [canvas.create_rectangle(*coords_rec,fill='#FCE69a')]
-        #coordinates of text relative to box upper left corner
-        coords_text = listadd(coords_rec, [3, 1 ,0 ,0])
-        canvas.showMember.append(canvas.create_text(coords_text[0], coords_text[1], anchor=tk.NW, text=member))
-        #resize box to fit text
-        coords_text = canvas.bbox(canvas.showMember[-1])
-        coords_rec = listadd(coords_text,[-3, 0, 3, 0])
-        canvas.coords(canvas.showMember[0],*coords_rec)
-    
-    elif event.type == tk.EventType['Leave']:
-        #remove hover field
-        try:
-            canvas.delete(*canvas.showMember)
-        #or do nothing if it is not iterable, i.e. is not set yet
-        except TypeError:
-            pass
-
-
-def selectBlock(event):
-    #move current block with mouse
-    canvas = event.widget
-    selected = set(canvas.find_overlapping(event.x-1, event.y-1, event.x+1, event.y+1))
-
-    #checks on selected area
-    # already selected by area?
-    area_select = set(canvas.find_withtag('selected'))
-    # click outside of selected area releases the select
-    if not selected.intersection(area_select):
-        canvas.dtag('all', 'selected')
-        for select in area_select:
-            if 'building' in canvas.gettags(select):
-                canvas.itemconfig(select, stipple='')
-    #else, clicked inside
-    else:
-        for ab in area_select:
-            #get correct floor for buildings in area select
-            if 'building' in canvas.gettags(ab):
-                building = canvas.getBuildingFromId(ab)
-                for id in building.id.values():
-                    canvas.addtag_withtag('selected', id)
-        #delete if erease mode on
-        if canvas.erase:
-            canvas.addtag_withtag('erase','selected')
-        else:
-            #move selected block
-            canvas.addtag_withtag('move','selected')
-            canvas.dtag('all', 'selected')
-
-    #if erase is on, erase block instead of painting or moving
-    erase = canvas.erase
-    
-    #only paint new building if mouse not over existing building and building button is active
-    paint_mode = canvas.master.active_button and not erase and not canvas.assign_mode
-    if not selected.intersection(set(canvas.find_withtag('building'))) and paint_mode:
-        #list makes it easier than working with int
-        selected = list()
-        #extend list if more than one object selected, else append value
-        nb = canvas.block.paint(canvas,[event.x,event.y],grid=False)
-        try:
-            #has to switch order, as building needs to be last
-            selected.extend(nb)
-        except TypeError:
-            selected.append(nb)
-        #only move buildings!
-    if selected: 
-        block = None
-        for select in selected:
-            if 'building' in canvas.gettags(select):
-                #canvas.new_block = list()
-                #canvas.new_block.append(select)
-                canvas.new_block = select
-                # remember building
-                new_building = canvas.getBuildingFromId(select)
-                #if erase is on, erase block instead of painting or moving
-                if erase:
-                    canvas.addtag('erase', 'withtag', select)
-                #if the selected building is a city and the assign mode is on, assign city to member
-                elif canvas.assign_mode and 'City' in canvas.gettags(select):    
-                    canvas.assignMember(ML=canvas.master.MembersList,city=select)
-                    # No selection in assign mode!
-                    canvas.dtag('all', 'selected')
-                else:
-                    if new_building is not None:
-                        for id in new_building.id.values():
-                            canvas.addtag_withtag('move', id)
-                            canvas.itemconfig(id,fill='blue')
-                    else:
-                        canvas.addtag_withtag('move', select)
-                        canvas.itemconfig(select,fill='blue')
-                canvas.startxy = (event.x, event.y)
-                #find type of building
-                #get first building type (should be only building type)
-                block = list(set(canvas.gettags(canvas.new_block)) & canvas.master.buildings)[0]
-                # Only take highest building, leave loop if found
-                break
-        
-        if erase:
-            eraseBlock(canvas)
-            return
-        if block is not None:
-            # define new block type
-            canvas.block = eval(f"{block}()")
-            canvas.block.canvas= canvas
-
-def eraseBlock(canvas):
-    # simply delete all blocks tagged 'erase'
-    blocks_to_erase = canvas.find_withtag('erase')
-    for block in blocks_to_erase:
-        canvas.removeBuilding(block)
-    #No block selected
-    canvas.new_block = None
-
-def moveBlock(event):
-    #move current block with mouse
-    canvas = event.widget
-    #if there is already a selected block, move it, else ignore it
-    move_blocks = canvas.find_withtag('move')
-    if len(move_blocks) != 0:
-        canvas.config(cursor='fleur')
-        try:
-            for new_block in move_blocks:
-                dx, dy = event.x-canvas.startxy[0], event.y-canvas.startxy[1]
-                canvas.move(new_block,dx,dy)   
-            canvas.startxy = (event.x, event.y)
-        except TypeError:
-            dx, dy = event.x-canvas.startxy[0], event.y-canvas.startxy[1]
-            canvas.move(move_blocks,dx,dy)   
-            canvas.startxy = (event.x, event.y)
-        #also show coordinates when moving
-        canvas.showCoords(event)
-
-def putBlock(event):
-    #paint the block
-    canvas = event.widget
-    # if canvas.new_block is not defined, set it to 'None'
-    try:
-        type(canvas.new_block)
-    except AttributeError:
-        canvas.new_block = None
-
-    # Do not put or delete block if in assign mode
-    if canvas.new_block is not None and not canvas.assign_mode:
-        #get the building from the building list
-        new_building = canvas.getBuildingFromId(canvas.new_block)
-        if isinstance(new_building,Block):
-            new_block = new_building.id['building']
-            #correct colors
-            for part in new_building.id.values():
-                if 'floor' in canvas.gettags(part):
-                    canvas.itemconfig(part,fill=new_building.floor_color)
-                elif 'member' in canvas.gettags(part):
-                    canvas.itemconfig(part,fill=used_colors["assign"])
-                else:
-                    canvas.itemconfig(part,fill=new_building.color,stipple='')
-                #remove 'move'status from new_block
-                canvas.dtag(part,'move')
-        else:
-            new_block = new_building
-            canvas.dtag(new_block,'move')
-        #take coordinates of center of building, NOT event coordinates!
-        ccoords = center(canvas.coords(new_block))
-
-        #check collision in canvas coords BEFORE applying grid!
-        delta_x, delta_y = checkCollision(canvas, new_block)
-        #change grid_coordinates if neccessary
-        
-        #calculate new coordinates
-        grid_coords = canvas.convCoord2Grid([ccoords[0]+delta_x, ccoords[1]+delta_y],block=canvas.block)
-        
-        #move building and check for collision on new location
-        #canvas.delete(*canvas.new_block)
-        canvas.moveBuilding(new_building, grid_coords, grid=True, rel=False)
-        #if it has a floor, take only the building
-        delta_x, delta_y = checkCollision(canvas, new_block)
-
-        # collect all positions the block was tried to put
-        grid_coords_hist = list()
-        while delta_x != 0 or delta_y != 0:
-            # check for further collisions
-            grid_coords_hist.append(grid_coords)
-            grid_coords = canvas.convCoord2Grid([ccoords[0]+delta_x, ccoords[1]+delta_y],block=canvas.block)
-            canvas.moveBuilding(new_building, grid_coords, grid=True, rel=False)
-            delta_x, delta_y = checkCollision(canvas, new_block)
-            # if position was already tried, break (otherwise inifinty loop!)
-            if grid_coords in grid_coords_hist:
-                # if the block is part of a bigger selected area, don't blink
-                if canvas.find_withtag('move'):
-                    break
-                else:
-                    canvas.itemconfig(new_block,fill='red')
-                    canvas.update()
-                    canvas.after(300)
-                    canvas.itemconfig(new_block,fill='blue')
-                #return
-                break
-        #get total delta of movement
-        delta_coord = listsub(center(canvas.coords(new_block)),ccoords)
-
-        # now move all selected blocks by the accumulated delta
-        for select in canvas.find_withtag('move'):
-            canvas.move(select,*delta_coord)
-            #remove stipple if it is not a floor
-            if 'floor' not in canvas.gettags(select):
-                canvas.itemconfig(select,stipple='')
-            canvas.dtag(select,'move')
-        # reset cursor
-        canvas.config(cursor='left_ptr')
-        # reset building type if active_button
-        if canvas.master.active_button:
-            if type(canvas.block) is type(canvas.active_block):
-                # There can be only one HQ and one Trap
-                if isinstance(canvas.block, HQ) or isinstance(canvas.block, Trap):
-                    canvas.master.activateButton(canvas.master.active_button)
-            else:
-                canvas.block = canvas.active_block
-        #remove 'moved' but should never be reached
-        for mb in canvas.find_withtag('move'):
-            canvas.delete(mb)
-            
-        canvas.new_block = None
-    #at the end, put assignment indicators on top
-    canvas.tag_raise('member','all')
-    #but still have the floor on top if so indicated
-    if canvas.top_floor:
-        canvas.tag_raise('floor','all')
-
-def checkCollision(canvas, block):
-    # check if new block collides with existing building
-    # tk rectangles give lower right corner OUTSIDE the rectangle, 
-    # so overlap area needs to be reduced to actual size
-    canv_box = listadd(canvas.coords(block),[1,1,-1,-1])
-    #canv_box.extend(block.convGrid2Coord(grid_box[2:4]))
-    #find list of all other blocks covering this area (except the new block itself)
-    selected = list(canvas.find_overlapping(*canv_box))
-    #block itself is always in the list, so remove it
-    selected.remove(block)
-    #grid coordinates block has to move
-    delta_coord = [0, 0]
-    for sb in selected:
-        if 'building' in canvas.gettags(sb): 
-            #get grid coordinates of sb rectangle, again reduce size of rectangle
-            #sb_box = listadd(canvas.coords(sb),[0,0,-1,-1])      
-            sb_box = canvas.coords(sb)      
-            delta_coord = pushBuilding(sb_box,canv_box)
-
-    return delta_coord[0], delta_coord[1]
-
-def pushBuilding(old_bb,new_bb):
-            #get deltas to move new_building outside ol_building with minimal distance
-            #old_bb: box of existing building in grid coordinates
-            #new_bb: box of new building in grid coordinates
-            #: make proper building class and methods
-
-            delta_coord = [0 , 0]
-            
-            #all buildings are squares!
-            #delta_size = abs(old_bb[2]-old_bb[0]) - abs(new_bb[2]-new_bb[0])
-            new_size = abs(new_bb[2]-new_bb[0])
-            old_size = abs(old_bb[2]-old_bb[0])
-            #x-axis:
-            #change if overlap in x-axis
-            if new_bb[0] < old_bb[2] and new_bb[2] > old_bb[0]:
-                delta_cx = center(new_bb)[0] - center(old_bb)[0] 
-
-                if delta_cx == 0:
-                    delta_coord[0] = abs(new_bb[2]-new_bb[0])
-                else:
-                    #delta_x = delta_x/abs(delta_x)*delta_size/2 -delta_x
-                    delta_coord[0] = (new_size+old_size)/2 - abs(delta_cx)
-                    #consider direction, apply sign
-                    delta_coord[0] *= delta_cx/abs(delta_cx)
-            #y-axis:
-            #change if overlap in y-axis
-            if new_bb[1] < old_bb[3] and new_bb[3] > old_bb[1]:
-                delta_cy = center(new_bb)[1] - center(old_bb)[1]
-                if delta_cy == 0:
-                    delta_coord[1] = abs(new_bb[3]-new_bb[1])
-                else:
-                    delta_coord[1] = (new_size+old_size)/2 - abs(delta_cy)
-                    #consider direction, apply sign
-                    delta_coord[1] *= delta_cy/abs(delta_cy)
-            #only one axis should be necessary unless it is completely covered/covering?
-            #use smaller absolute delta (only if both axos have a delta)
-            #is there a 0 velue in delta_coordss, then leave it
-            if 0 not in delta_coord:
-                delta_coord[delta_coord.index(max(delta_coord,key=abs))] = 0
-
-            return delta_coord
+        #at the end center the view at the same point as in the PaintCanvas
+        upper_canvas.xview_moveto(self.upper_canvas.xview()[0])
+        upper_canvas.yview_moveto(self.upper_canvas.yview()[0])
 
 def make_png(widget, toplevel):
     save_file = asksaveasfilename(title='Save Hive as PNG:', defaultextension='.png', initialdir=save_dir)
@@ -1863,3 +1374,4 @@ def make_png(widget, toplevel):
 
 MW = MainWindow()
 MW.mainloop()
+MW.destroy()
